Enhanced Readme of R-Recursion
